feat(os): GPU render + WM integration tests — 43/43 pass
GPU 2D render test (gpu_render_test_entry.spl, 16/16):
  - Solid fill + pixel read-back (red R=255, G=0, B=0)
  - Colored rectangles (red, green, blue, white — each verified)
  - Alpha blending (white + 50% red → mixed channels verified)
  - Rect boundary precision (inside green, outside black)
  - Rect overwrite (green on top of red = green)

WM integration test (wm_test_entry.spl, 27/27):
  - Open Hello World window: title bar, close button, body, background verified
  - Type text: text indicator rendered in window body
  - Drag window: old position=background, new position=title bar
  - Maximize: fills 1024x768, corners verified
  - Minimize: window hidden, screen all background
  - Restore: window visible, geometry restored to pre-maximize
  - Close: window removed, background verified

Fix: rt_gui_read_pixel returns raw u32 (not tagged RuntimeValue)
Fix: use rt_gui_blend_fill for shadow buffer writes (rt_gui_fill4 is MMIO-only)

+++ examples/simple_os/arch/x86_64/boot/glass_render.c
@@ -618,7 +618,7 @@ RuntimeValue rt_gui_read_pixel(RuntimeValue x_y, RuntimeValue u1,
     uint32_t x = (uint32_t)((uint64_t)x_y >> 32);
     uint32_t y = (uint32_t)((uint64_t)x_y & 0xFFFFFFFF);
     if (x >= g_fb_w || y >= SCREEN_H) return 0;
-    return (RuntimeValue)fb_read(x, y);
+    return (RuntimeValue)(uint64_t)fb_read(x, y);
 }
 
 /* ===================================================================
